Clear expired Django sessions regularly from DB
A random minute in the 21st hour of Sunday UTC was chosen arbitrarily.
That it is run on a single arbitrary host from the set was done to
prevent unnecessary work being done until Bug #1424705 provides us with
a cleaner solution.

The bug as filed seems to indicate that this is not something that
needs to be done daily in high-use environments so a standard of once
per week was chosen. Making this configurable or tunable does not seem
to be necessary at this time.

Change-Id: I87164926a4d6a65d3c9f733d625a2d4af91c2597
Closes-Bug: #1466126